Clean up int problems on i386 and methods problems

+++ NumToWords/EntryFr.go
@@ -29,7 +29,7 @@ func (Entry EntryFr) Translate(Input int) string {
 		result = Entry.handleThousands(Input)
 	case Input < 1000000000:
 		result = Entry.handleMillions(Input)
-	case Input < 1000000000000:
+	case int64(Input) < 1000000000000:
 		result = Entry.handleBillions(Input)
 	default:
 		result = Entry.handleTrillions(Input)
@@ -150,8 +150,8 @@ func (Entry EntryFr) handleBillions(Input int) string {
 }
 
 func (Entry EntryFr) handleTrillions(Input int) string {
-	trillions := Input / 1000000000000
-	remainder := Input % 1000000000000
+	trillions := int(int64(Input) / 1000000000000)
+	remainder := int(int64(Input) % 1000000000000)
 
 	plural := "billion"
 	if trillions > 1 {
